Refactor form page into dynamic route
