--- data/junior1 copy.json
@@ -1,279 +1,0 @@
-{
-    "questions": [
-        {
-            "id": 1,
-            "question": "Какой тип данных используется для хранения целых чисел в JavaScript?",
-            "answers": {
-                "a": "Number",
-                "b": "Integer",
-                "c": "Float",
-                "d": "Decimal"
-            },
-            "correct": [
-                "a", "b"
-            ]
-        },
-        {
-            "id": 2,
-            "question": "Какой тег HTML используется для создания гиперссылки?",
-            "answers": [
-                "<a>",
-                "<link>",
-                "<href>",
-                "<hyperlink>"
-            ],
-            "correct": [
-                "<a>"
-            ]
-        },
-        {
-            "id": 3,
-            "question": "Какие из следующих языков являются языками программирования?",
-            "checkbox": true,
-            "answers": [
-                "Python",
-                "HTML",
-                "Java",
-                "CSS"
-            ],
-            "correct": [
-                "Python",
-                "Java"
-            ]
-        },
-        {
-            "id": 4,
-            "question": "Какой протокол используется для безопасной передачи данных по сети?",
-            "answers": [
-                "HTTP",
-                "FTP",
-                "SMTP",
-                "HTTPS"
-            ],
-            "correct": [
-                "HTTPS"
-            ]
-        },
-        {
-            "id": 5,
-            "question": "Что означает аббревиатура CPU?",
-            "answers": [
-                "Central Processing Unit",
-                "Central Programming Unit",
-                "Computer Processing Unit",
-                "Control Processing Unit"
-            ],
-            "correct": [
-                "Central Processing Unit"
-            ]
-        },
-        {
-            "id": 6,
-            "question": "Какие из этих баз данных являются реляционными?",
-            "checkbox": true,
-            "answers": [
-                "MongoDB",
-                "PostgreSQL",
-                "MySQL",
-                "Redis"
-            ],
-            "correct": [
-                "PostgreSQL",
-                "MySQL"
-            ]
-        },
-        {
-            "id": 7,
-            "question": "Какой язык используется для стилизации веб-страниц?",
-            "answers": [
-                "HTML",
-                "Python",
-                "CSS",
-                "SQL"
-            ],
-            "correct": [
-                "CSS"
-            ]
-        },
-        {
-            "id": 8,
-            "question": "Что такое Git?",
-            "answers": [
-                "Система управления версиями",
-                "Язык программирования",
-                "База данных",
-                "Операционная система"
-            ],
-            "correct": [
-                "Система управления версиями"
-            ]
-        },
-        {
-            "id": 9,
-            "question": "Какие HTTP-методы используются для изменения данных?",
-            "checkbox": true,
-            "answers": [
-                "GET",
-                "POST",
-                "PUT",
-                "DELETE"
-            ],
-            "correct": [
-                "POST",
-                "PUT",
-                "DELETE"
-            ]
-        },
-        {
-            "id": 10,
-            "question": "Что такое DOM в контексте веб-разработки?",
-            "answers": [
-                "Document Object Model",
-                "Digital Object Mapping",
-                "Data Organization Method",
-                "Document Output Management"
-            ],
-            "correct": [
-                "Document Object Model"
-            ]
-        },
-        {
-            "id": 11,
-            "question": "Какая команда используется для клонирования репозитория в Git?",
-            "answers": [
-                "git copy",
-                "git download",
-                "git clone",
-                "git fork"
-            ],
-            "correct": [
-                "git clone"
-            ]
-        },
-        {
-            "id": 12,
-            "question": "Какие из этих языков компилируются?",
-            "checkbox": true,
-            "answers": [
-                "Java",
-                "Python",
-                "C++",
-                "JavaScript"
-            ],
-            "correct": [
-                "Java",
-                "C++"
-            ]
-        },
-        {
-            "id": 13,
-            "question": "Какой порт по умолчанию используется для HTTPS?",
-            "answers": [
-                "80",
-                "21",
-                "443",
-                "22"
-            ],
-            "correct": [
-                "443"
-            ]
-        },
-        {
-            "id": 14,
-            "question": "Что такое API?",
-            "answers": [
-                "Асинхронный процессор интерфейса",
-                "Протокол взаимодействия баз данных",
-                "Интерфейс программирования приложений",
-                "Модель данных веб-приложения"
-            ],
-            "correct": [
-                "Интерфейс программирования приложений"
-            ]
-        },
-        {
-            "id": 15,
-            "question": "Какие из этих форматов являются форматами изображений?",
-            "checkbox": true,
-            "answers": [
-                "PNG",
-                "JSON",
-                "SVG",
-                "JPG"
-            ],
-            "correct": [
-                "PNG",
-                "SVG",
-                "JPG"
-            ]
-        },
-        {
-            "id": 16,
-            "question": "Что делает оператор === в JavaScript?",
-            "answers": [
-                "Сравнивает значения без учета типа",
-                "Присваивает значение переменной",
-                "Сравнивает значения и типы",
-                "Проверяет, является ли значение истиной"
-            ],
-            "correct": [
-                "Сравнивает значения и типы"
-            ]
-        },
-        {
-            "id": 17,
-            "question": "Какой метод используется для вывода данных в консоль в JavaScript?",
-            "answers": [
-                "print()",
-                "echo()",
-                "console.log()",
-                "write()"
-            ],
-            "correct": [
-                "console.log()"
-            ]
-        },
-        {
-            "id": 18,
-            "question": "Какие из следующих являются типами данных в Python?",
-            "checkbox": true,
-            "answers": [
-                "int",
-                "list",
-                "object",
-                "map"
-            ],
-            "correct": [
-                "int",
-                "list",
-                "object"
-            ]
-        },
-        {
-            "id": 19,
-            "question": "Какой язык программирования используется в Android-приложениях?",
-            "answers": [
-                "Swift",
-                "Kotlin",
-                "Ruby",
-                "PHP"
-            ],
-            "correct": [
-                "Kotlin"
-            ]
-        },
-        {
-            "id": 20,
-            "question": "Какой из этих протоколов работает на канальном уровне модели OSI?",
-            "answers": [
-                "IP",
-                "TCP",
-                "HTTP",
-                "Ethernet"
-            ],
-            "correct": [
-                "Ethernet"
-            ]
-        }
-    ]
-}
