--- data/juniour.json
@@ -1,228 +1,0 @@
-{
-  "metadata": {
-    "title": "Основы программирования и IT",
-    "status": "Update"
-  },
-  "questions": [
-    {
-      "id": 1,
-      "question": "Какой тип данных используется для хранения целых чисел в JavaScript?",
-      "answers": {
-        "a": "Number",
-        "b": "Integer",
-        "c": "Float",
-        "d": "Decimal"
-      },
-      "correct": ["a"]
-    },
-    {
-      "id": 2,
-      "question": "Как обозначается функция в JavaScript?",
-      "answers": {
-        "a": "func",
-        "b": "function",
-        "c": "def",
-        "d": "fn"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 3,
-      "question": "Что такое HTML?",
-      "answers": {
-        "a": "Язык программирования",
-        "b": "Фреймворк",
-        "c": "Язык разметки",
-        "d": "СУБД"
-      },
-      "correct": ["c"]
-    },
-    {
-      "id": 4,
-      "question": "Какой метод используется для вывода в консоль в JavaScript?",
-      "answers": {
-        "a": "print()",
-        "b": "log()",
-        "c": "console.log()",
-        "d": "write()"
-      },
-      "correct": ["c"]
-    },
-    {
-      "id": 5,
-      "question": "Что такое CSS?",
-      "answers": {
-        "a": "Язык запросов",
-        "b": "Каскадные таблицы стилей",
-        "c": "Скриптовый язык",
-        "d": "Протокол передачи данных"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 6,
-      "question": "Как объявить переменную в JavaScript?",
-      "answers": {
-        "a": "var",
-        "b": "const",
-        "c": "let",
-        "d": "define"
-      },
-      "correct": ["a", "b", "c"]
-    },
-    {
-      "id": 7,
-      "question": "Что такое API?",
-      "answers": {
-        "a": "Протокол маршрутизации",
-        "b": "Интерфейс взаимодействия программ",
-        "c": "База данных",
-        "d": "Язык программирования"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 8,
-      "question": "Какой из следующих вариантов — фреймворк JavaScript?",
-      "answers": {
-        "a": "Vue",
-        "b": "React",
-        "c": "Angular",
-        "d": "Laravel"
-      },
-      "correct": ["a", "b", "c"]
-    },
-    {
-      "id": 9,
-      "question": "Что возвращает оператор typeof?",
-      "answers": {
-        "a": "Тип переменной",
-        "b": "Значение переменной",
-        "c": "Длину строки",
-        "d": "Название функции"
-      },
-      "correct": ["a"]
-    },
-    {
-      "id": 10,
-      "question": "Какой тег используется для заголовка первого уровня в HTML?",
-      "answers": {
-        "a": "<head>",
-        "b": "<h1>",
-        "c": "<title>",
-        "d": "<header>"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 11,
-      "question": "Что такое DOM?",
-      "answers": {
-        "a": "Объектная модель документа",
-        "b": "Массив данных",
-        "c": "Протокол передачи",
-        "d": "Сервис хранения"
-      },
-      "correct": ["a"]
-    },
-    {
-      "id": 12,
-      "question": "Какой протокол используется для обмена данными по сети?",
-      "answers": {
-        "a": "HTTP",
-        "b": "FTP",
-        "c": "SSH",
-        "d": "UDP"
-      },
-      "correct": ["a"]
-    },
-    {
-      "id": 13,
-      "question": "Что делает метод Array.push() в JavaScript?",
-      "answers": {
-        "a": "Удаляет элемент",
-        "b": "Добавляет элемент в конец массива",
-        "c": "Сортирует массив",
-        "d": "Изменяет объект"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 14,
-      "question": "Какие значения в JavaScript считаются ложными (falsy)?",
-      "answers": {
-        "a": "0",
-        "b": "null",
-        "c": "undefined",
-        "d": "true"
-      },
-      "correct": ["a", "b", "c"]
-    },
-    {
-      "id": 15,
-      "question": "Что такое JSON?",
-      "answers": {
-        "a": "Язык программирования",
-        "b": "Формат данных",
-        "c": "Фреймворк",
-        "d": "Протокол"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 16,
-      "question": "Какой оператор используется для строгого сравнения?",
-      "answers": {
-        "a": "==",
-        "b": "=",
-        "c": "===",
-        "d": "!="
-      },
-      "correct": ["c"]
-    },
-    {
-      "id": 17,
-      "question": "Как задать комментарий в CSS?",
-      "answers": {
-        "a": "// комментарий",
-        "b": "/* комментарий */",
-        "c": "# комментарий",
-        "d": "<!-- комментарий -->"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 18,
-      "question": "Как вызвать всплывающее окно в JavaScript?",
-      "answers": {
-        "a": "prompt()",
-        "b": "alert()",
-        "c": "confirm()",
-        "d": "popup()"
-      },
-      "correct": ["b"]
-    },
-    {
-      "id": 19,
-      "question": "Что такое GIT?",
-      "answers": {
-        "a": "Система контроля версий",
-        "b": "Язык программирования",
-        "c": "Текстовый редактор",
-        "d": "База данных"
-      },
-      "correct": ["a"]
-    },
-    {
-      "id": 20,
-      "question": "Какая команда создаёт новый репозиторий Git?",
-      "answers": {
-        "a": "git init",
-        "b": "git start",
-        "c": "git create",
-        "d": "git repo"
-      },
-      "correct": ["a"]
-    }
-  ]
-}
