+++ data/new.json
@@ -1,0 +1,75 @@
+{
+  "title": "User Registration Form",
+  "description": "Please fill out all required fields to create your account.",
+  "fields": [
+    {
+      "id": 1,
+      "name": "interests",
+      "label": "Interests",
+      "type": "radioButton",
+      "question": "Interests",
+      "required": false,
+      "options": [
+        { "label": "Technology", "value": "tech" },
+        { "label": "Sports", "value": "sports" },
+        { "label": "Music", "value": "music" },
+        { "label": "Travel", "value": "travel" },
+        { "label": "Cooking", "value": "cooking" },
+        { "label": "Reading", "value": "reading" }
+      ]
+    },
+    {
+      "id": 2,
+      "name": "newsletter",
+      "label": "Subscribe to Newsletter",
+      "type": "checkboxGroup",
+      "question": "Subscribe to Newsletter",
+      "required": false,
+      "options": [
+        { "label": "some q", "value": "dfs43" },
+        { "label": "some q2", "value": "dfs62" },
+        { "label": "some q3", "value": "dfs16" }
+      ]
+    },
+    {
+      "id": 3,
+      "name": "quizQuestion1",
+      "label": "What is your preferred programming language?",
+      "type": "checkboxGroup",
+      "question": "What is your preferred programming language?",
+      "required": true,
+      "options": [
+        { "label": "JavaScript/TypeScript", "value": "js" },
+        { "label": "Python", "value": "python" },
+        { "label": "Java", "value": "java" },
+        { "label": "C++", "value": "cpp" },
+        { "label": "Go", "value": "go" },
+        { "label": "Rust", "value": "rust" }
+      ]
+    },
+    {
+      "id": 4,
+      "name": "quizQuestion2",
+      "label": "Which development methodologies do you use?",
+      "type": "checkboxGroup",
+      "question": "Which development methodologies do you use?",
+      "required": false,
+      "options": [
+        { "label": "Agile/Scrum", "value": "agile" },
+        { "label": "Waterfall", "value": "waterfall" },
+        { "label": "Kanban", "value": "kanban" },
+        { "label": "DevOps", "value": "devops" },
+        { "label": "Test-Driven Development", "value": "tdd" },
+        { "label": "Behavior-Driven Development", "value": "bdd" }
+      ]
+    },
+    {
+      "id": 5,
+      "name": "feedback",
+      "label": "feedback",
+      "type": "textarea",
+      "question": "Enter you feedback ",
+      "required": false
+    }
+  ]
+}
